fix(ci): keep ES5 source-map printer under guard
